+++ CHANGELOG.md
@@ -1,14 +1,18 @@
-## 0.1.2-dev1
+# 0.2.0-dev0
 
 ### Enhancements
 
 * **Added migration for GitHub Source V2**
   * **Updated the expected output for GitHub Source V2**
 
-## 0.1.2-dev0
+## 0.2.0
 
 ### Enhancements
 
+* **Add snowflake source and destination connectors**
+* **Migrate Slack Source Connector to V2**
+* **Migrate Slack Source Connector to V2**
+* **Add Delta Table destination to v2**
 * **Migrate Slack Source Connector to V2**
 
 ## 0.1.1
@@ -18,6 +22,10 @@
 * **Update KDB.AI vectorstore integration to 1.4**
 * **Add sqlite and postgres source connectors**
 * **Add sampling functionality for indexers in fsspec connectors**
+
+### Fixes
+
+* **Fix Databricks Volumes destination** Fix for filenames to not be hashes.
 
 ## 0.1.0